+++ src/main/scala/onion/compiler/typing/ClosureTyping.scala
@@ -488,7 +488,7 @@ final class ClosureTyping(
   }
 
   private def leastUpperBound(node: AST.Node, left: Type, right: Type): Type =
-    TypeCheckingHelpers.leastUpperBound(node, left, right, bodyContext.rootClass,
+    TypeCheckingHelpers.leastUpperBound(bodyContext.table, node, left, right, bodyContext.rootClass,
       (n, l, r) => bodyContext.report(INCOMPATIBLE_TYPE, n, l, r))
 
   private def buildReturnBlock(node: AST.BlockExpression, bodyTerm: Term, returnType: Type): ActionStatement = {